+++ Calculator/CalculatorTests/ExpressionParserTests.swift
@@ -7,24 +7,6 @@
 
 import XCTest
 
-
-enum ExpressionParser {
-    static func parse(from input: String) -> Formula {
-        let operands = componentsByOperators(from: input)
-                        .compactMap{ Double($0) }
-        let operators = input.map{ $0 }
-                        .compactMap{ Operator(rawValue: $0) }
-        return Formula(operands: CalculatorItemQueue(operands),
-                       operators: CalculatorItemQueue(operators))
-    }
-    
-    private static func componentsByOperators(from input: String) -> [String] {
-        let sliceInputValue = input.split()
-        let operands = sliceInputValue.filter { Double($0.description) != nil }
-        return operands
-    }
-}
-
 class ExpressionParserTests: XCTestCase {
 
     func test_연산자와숫자가_한개씩() {
